# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /me:
    get:
      description: Retrieves the profile of the authenticated user
      operationId: UsersController_getProfile
      parameters: []
      responses:
        '200':
          description: Profile retrieved successfully
        '401':
          description: Unauthorized
      security:
        - bearer: []
      summary: Get current user profile
      tags:
        - me
    patch:
      description: Updates the profile of the authenticated user
      operationId: UsersController_updateProfile
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileDto'
      responses:
        '200':
          description: Profile updated successfully
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
      security:
        - bearer: []
      summary: Update current user profile
      tags:
        - me
...
  /admin/users:
    get:
      description: Retrieves all users in the system. Requires admin or sysadmin role.
      operationId: AdminController_getAllUsers
      parameters: []
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponseDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires admin or sysadmin role
      security:
        - bearer: []
      summary: Get all users
      tags:
        - admin
  /admin/users/{id}:
    get:
      description: Retrieves a specific user by their ID. Requires admin or sysadmin role.
      operationId: AdminController_getUserById
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires admin or sysadmin role
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Get user by ID
      tags:
        - admin
    patch:
      description: >-
        Updates a user by their ID. Can update email, password, full name, role,
        and email verification status. Requires sysadmin role.
      operationId: AdminController_updateUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Invalid input data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires sysadmin role
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Update user
      tags:
        - admin
    delete:
      description: Deletes a user by their ID. Requires sysadmin role.
      operationId: AdminController_deleteUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires sysadmin role
        '404':
          description: User not found
      security:
        - bearer: []
      summary: Delete user
      tags:
        - admin
  /admin/merge-users:
    post:
      description: >-
        Merges source user account into destination user account. Source account
        will be deleted after merge. Fails if there are overlapping identities.
        Requires sysadmin role.
      operationId: AdminController_mergeAccounts
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeAccountsDto'
      responses:
        '200':
          description: Accounts merged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MergeAccountsResponseDto'
        '400':
          description: Cannot merge - overlapping identities or invalid request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Requires sysadmin role
        '404':
          description: Source or destination user not found
      security:
        - bearer: []
      summary: Merge user accounts
      tags:
        - admin
...
info:
  title: Todo App API
  description: >-
    REST API for Todo Application with authentication and role-based access
    control
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    UpdateProfileDto:
      type: object
      properties:
        fullName:
          type: string
          description: User full name
          example: Jane Smith
          minLength: 1
          maxLength: 255
...
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          description: User ID
          example: uuid-123
        email:
          type: string
          description: User email address (primary)
          example: user@example.com
        fullName:
          type: string
          description: User full name
          example: John Doe
        role:
          type: string
          description: User role
          example: guest
        emailVerifiedAt:
          type: object
          description: Email verification timestamp
          example: '2025-01-01T00:00:00.000Z'
          nullable: true
        localUsername:
          type: object
          description: Local username (could be email or username)
          example: john.doe
          nullable: true
        googleEmail:
          type: object
          description: Google account email
          example: user@gmail.com
          nullable: true
        msEmail:
          type: object
          description: Microsoft account email
          example: user@outlook.com
          nullable: true
        createdAt:
          format: date-time
          type: string
          description: User creation timestamp
          example: '2025-01-01T00:00:00.000Z'
        updatedAt:
          format: date-time
          type: string
          description: User last update timestamp
          example: '2025-01-01T00:00:00.000Z'
      required:
        - id
        - email
        - fullName
        - role
        - emailVerifiedAt
        - localUsername
        - googleEmail
        - msEmail
        - createdAt
        - updatedAt
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          description: User email address
          example: user@example.com
        password:
          type: string
          description: User password (minimum 8 characters)
          example: newPassword123
          minLength: 8
          maxLength: 128
        fullName:
          type: string
          description: User full name
          example: John Doe
          minLength: 1
          maxLength: 255
        role:
          type: string
          description: User role
          enum:
            - guest
            - admin
            - sysadmin
          example: guest
        emailVerifiedAt:
          type: object
          description: Email verification timestamp (ISO 8601 format, or null to unverify)
          example: '2025-01-01T00:00:00.000Z'
    MergeAccountsDto:
      type: object
      properties:
        sourceUserId:
          type: string
          description: Source user ID (will be deleted after merge)
          example: 550e8400-e29b-41d4-a716-446655440000
        destinationUserId:
          type: string
          description: Destination user ID (will receive merged identities)
          example: 660e8400-e29b-41d4-a716-446655440111
      required:
        - sourceUserId
        - destinationUserId
    MergeAccountsResponseDto:
      type: object
      properties:
        message:
          type: string
          description: Success message
        destinationUserId:
          type: string
          description: Destination user ID
        mergedIdentities:
          type: object
          description: Identities merged from source
      required:
        - message
        - destinationUserId
        - mergedIdentities